+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,20 +1,29 @@
 package org.apache.coyote.http11;
 
 import java.io.BufferedReader;
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.Socket;
+import java.net.URISyntaxException;
+import java.net.URL;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import nextstep.jwp.controller.FrontController;
+import java.util.Optional;
+import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UncheckedServletException;
+import nextstep.jwp.model.User;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.ContentType;
 import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.response.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -40,10 +49,8 @@ public class Http11Processor implements Runnable, Processor {
 
             final List<String> request = getRequest(inputStream);
             final HttpRequest httpRequest = new HttpRequest(request);
-            log.debug(httpRequest.toString());
 
-            FrontController controller = new FrontController();
-            HttpResponse httpResponse = controller.run(httpRequest);
+            HttpResponse httpResponse = getResponse(httpRequest);
             String response = httpResponse.parseToString();
             outputStream.write(response.getBytes());
             outputStream.flush();
@@ -67,5 +74,44 @@ public class Http11Processor implements Runnable, Processor {
             line = reader.readLine();
         }
         return request;
+    }
+
+    public HttpResponse getResponse(HttpRequest httpRequest) throws IOException {
+        if (httpRequest.getRequestUri().equals("/")) {
+            final String responseBody = "Hello world!";
+            return new HttpResponse(httpRequest.getProtocol(), HttpStatus.OK, ContentType.HTML, responseBody);
+        }
+        if (httpRequest.getRequestUri().endsWith(".html")) {
+            if (httpRequest.getRequestUri().equals("/login.html")) {
+                Optional<User> user = InMemoryUserRepository.findByAccount(
+                        httpRequest.getQueryStrings().get("account"));
+                user.ifPresent(value -> log.debug(value.toString()));
+            }
+            final String responseBody = getResponseBody(httpRequest);
+            return new HttpResponse(httpRequest.getProtocol(), HttpStatus.OK, ContentType.HTML, responseBody);
+        }
+        if (httpRequest.getRequestUri().endsWith(".css")) {
+            final String responseBody = getResponseBody(httpRequest);
+            return new HttpResponse(httpRequest.getProtocol(), HttpStatus.OK, ContentType.CSS, responseBody);
+
+        }
+        if (httpRequest.getRequestUri().endsWith(".js")) {
+            final String responseBody = getResponseBody(httpRequest);
+            return new HttpResponse(httpRequest.getProtocol(), HttpStatus.OK, ContentType.JAVASCRIPT,
+                    responseBody);
+        }
+        return new HttpResponse(httpRequest.getProtocol(), HttpStatus.NOT_FOUND, ContentType.HTML, null);
+    }
+
+    private String getResponseBody(final HttpRequest httpRequest) throws IOException {
+        final URL resourceURL = getClass().getClassLoader()
+                .getResource("static" + httpRequest.getRequestUri());
+        try {
+            final File file = new File(Objects.requireNonNull(resourceURL).toURI());
+            final Path path = file.getAbsoluteFile().toPath();
+            return new String(Files.readAllBytes(path));
+        } catch (URISyntaxException e) {
+            throw new RuntimeException(e);
+        }
     }
 }
